Move peer status handlers to it's own file
